backend/requirements.txt
aiohttp==3.9.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
Django==4.1.13
django-allauth==0.63.2
django-appconf==1.0.6
django-compressor==4.4
django-cors-headers==4.3.1
django-extensions==3.2.3
django-sslserver==0.22
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djongo==1.3.6
dnspython==2.6.1
dotenv==0.0.5
<<<<<<< Updated upstream
install==1.3.5
MarkupSafe==2.1.5
mongoengine==0.28.2
=======
frozenlist==1.4.1
idna==3.7
importlib-metadata==7.1.0
install==1.3.5
Markdown==3.6
MarkupSafe==2.1.5
mongoengine==0.28.2
multidict==6.0.5
>>>>>>> Stashed changes
numpy==1.26.4
pandas==2.2.2
psycopg2-binary==2.9.9
pycparser==2.22
pycryptodome==3.20.0
PyJWT==2.8.0
pymongo==4.7.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
rcssmin==1.1.1
requests==2.32.3
rjsmin==1.2.1
six==1.16.0
sqlparse==0.2.4
twilio==9.2.0
typing-extensions==4.12.0
tzdata==2024.1
<<<<<<< Updated upstream
=======
urllib3==2.2.2
>>>>>>> Stashed changes
werkzeug==3.0.3
whitenoise==6.6.0
yarl==1.9.4
zipp==3.19.2
